+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* ``Chunk`` function added.
 
 .. 1.0.0 (yyyy-mm-dd)
 .. ------------------